+++ geopandas/geodataframe.py
@@ -180,7 +180,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         return geopandas.io.file.read_file(filename, **kwargs)
 
     @classmethod
-    def from_features(cls, features, crs=None):
+    def from_features(cls, features, crs=None, columns=None):
         """
         Alternate constructor to create GeoDataFrame from an iterable of
         features or a feature collection.
@@ -196,6 +196,10 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
               ``__geo_interface__``.
         crs : str or dict (optional)
             Coordinate reference system to set on the resulting frame.
+        columns : list of column names, optional
+            Optionally specify the column names to include in the output frame.
+            This does not overwrite the property names of the input, but can
+            ensure a consistent output format.
 
         Returns
         -------
@@ -228,7 +232,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
             d = {'geometry': shape(f['geometry']) if f['geometry'] else None}
             d.update(f['properties'])
             rows.append(d)
-        df = GeoDataFrame.from_dict(rows)
+        df = GeoDataFrame(rows, columns=columns)
         df.crs = crs
         return df